правки по верстке, имя игрока поменял с имени на логин, проверка при сохранении по логин
