+++ packages/client/src/pages/error/Error500.tsx
@@ -1,17 +1,12 @@
-import { useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
-import { FORM_PAGE_CONTAINER_CLASS } from '../../constants/style-groups';
-import { IconButton } from '../../components/IconButton';
-import { EIconButton } from '../../enums';
-import { type PageInitArgs } from '../../routes';
+import {
+  APP_TITLE_CLASS,
+  FORM_PAGE_CONTAINER_CLASS,
+} from '../../constants/style-groups';
+import { type PageInitArgs, ROUTES } from '../../routes';
 
 export const Error500 = () => {
-  const navigate = useNavigate();
-
-  const goBack = () => {
-    navigate(-1);
-  };
-
   return (
     <>
       <Helmet>
@@ -19,14 +14,12 @@ export const Error500 = () => {
         <title>Страница ошибки</title>
         <meta name="description" content="Страница ошибки" />
       </Helmet>
-      <div className={FORM_PAGE_CONTAINER_CLASS}>
-        <div className="absolute top-6 left-6">
-          <IconButton
-            onClick={goBack}
-            iconName={EIconButton.BACK}
-            hoverName={'Назад'}
-          />
-        </div>
+      <div className={`${FORM_PAGE_CONTAINER_CLASS} flex-col`}>
+        <Link
+          to={ROUTES.main}
+          className={`${APP_TITLE_CLASS} mb-[-50px] no-underline z-10`}>
+          Перейти на главную страницу
+        </Link>
         <img
           className="max-w-full max-h-[70vh] w-auto h-auto object-contain"
           src={'src/assets/images/500.png'}
