+++ packages/client/src/pages/game/GameFinishScreen.tsx
@@ -1,4 +1,5 @@
 import { type FC } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { type GameResult } from './types';
 import {
   APP_TITLE_CLASS,
@@ -11,37 +12,39 @@ import { Button } from '../../components/Button';
 import { useDispatch, useSelector } from '../../store/store';
 import { selectUser, selectUserRating } from '../../slices/user-slice';
 import { updateLeaderboardScore } from '../../slices/leaderboard-slice';
+import { ROUTES } from '../../routes';
 
 interface GameFinishScreenProps {
   result: GameResult;
   onRestart: () => void;
 }
 
-export const GameFinishScreen: FC<GameFinishScreenProps> = ({ result, onRestart }) => {
+export const GameFinishScreen: FC<GameFinishScreenProps> = ({
+  result,
+  onRestart,
+}) => {
   const dispatch = useDispatch();
   const user = useSelector(selectUser);
+  const navigate = useNavigate();
 
-
-  if(user && result.winnerName === user?.first_name){
+  if (user && result.winnerName === user?.login) {
     const userRating = useSelector(selectUserRating);
     const newScore = userRating + result.pointGame;
 
     const handleRecordRating = () => {
       dispatch(
         updateLeaderboardScore({
-          data: 
-            {
-              id: user.id ?? '',
-              name: result.winnerName,
-              flip7_rating: newScore,
-            }
+          data: {
+            id: user.id ?? '',
+            name: result.winnerName,
+            flip7_rating: newScore,
+          },
         })
       );
     };
 
     handleRecordRating();
   }
-  
 
   return (
     <div className={FORM_PAGE_CONTAINER_CLASS}>
@@ -61,6 +64,7 @@ export const GameFinishScreen: FC<GameFinishScreenProps> = ({ result, onRestart 
         </ul>
         <div className={BTN_GROUP_CLASS}>
           <Button onClick={onRestart} content="Играть снова" />
+          <Button onClick={() => navigate(ROUTES.main)} content="Выйти" />
         </div>
       </div>
     </div>
