+++ packages/client/src/pages/game/GamePage.tsx
@@ -1,17 +1,10 @@
 import { type FC, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
 import { DEFAULT_GAME_CONFIG, type GameConfig, type GameResult } from './types';
 import { GameStartScreen } from './GameStartScreen';
 import { GamePlayScreen } from './GamePlayScreen';
 import { GameFinishScreen } from './GameFinishScreen';
-import {
-  GAME_HEADER_CONTAINER_CLASS,
-  GAME_HEADER_DIV_ITEMS_CLASS,
-  ITEMS_CENTER_CLASS,
-} from '../../constants/style-groups';
-import { Button } from '../../components/Button';
-import { type PageInitArgs, ROUTES } from '../../routes';
+import { type PageInitArgs } from '../../routes';
 
 enum GameState {
   Start = 'start',
@@ -23,8 +16,6 @@ export const GamePage: FC = () => {
   const [gameState, setGameState] = useState<GameState>(GameState.Start);
   const [gameConfig, setGameConfig] = useState<GameConfig>(DEFAULT_GAME_CONFIG);
   const [gameResult, setGameResult] = useState<GameResult | null>(null);
-
-  const navigate = useNavigate();
 
   const handleStart = (config: GameConfig) => {
     setGameConfig(config);
@@ -50,34 +41,6 @@ export const GamePage: FC = () => {
       {gameState === GameState.Start && (
         <GameStartScreen onStart={handleStart} />
       )}
-      <header
-        className={`${GAME_HEADER_CONTAINER_CLASS} ${ITEMS_CENTER_CLASS}`}>
-        <div className={`${GAME_HEADER_DIV_ITEMS_CLASS}`}>
-          <Button
-            onClick={() => navigate(ROUTES.main)}
-            content={
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="24"
-                height="24"
-                viewBox="0 0 24 24">
-                <title>Arrow-back SVG Icon</title>
-                <path
-                  className="fill-path-dark"
-                  d="M21 11H6.414l5.293-5.293l-1.414-1.414L2.586 12l7.707 7.707l1.414-1.414L6.414 13H21z"
-                />
-              </svg>
-            }
-          />
-        </div>
-        <div className={`${GAME_HEADER_DIV_ITEMS_CLASS}`}>
-          <Button
-            onClick={() => window.location.reload()}
-            content={'Завершить игру'}
-          />
-        </div>
-        <div></div>
-      </header>
       {gameState === GameState.Playing && (
         <GamePlayScreen config={gameConfig} onFinish={handleFinish} />
       )}
